# Faster disjoint/isSubsetOf for Set via unbalanced splitting.