@alexfmpe 2022-10-29
The same should be applicable to `Map`, but there `disjoint` and `isSubmapOf` use different splits, resp, `splitMember` and `splitLookup`. Do I duplicate both to replace `link` with `bin` ?
FWIW, replacing `splitMember` with `splitMember k0 m = let (a,b,c) = splitLookup k0 m in (a, isJust b, c)` seems to cause no change in the benchmarks, but I don't know how fragile the optimizations involved are.