@alexfmpe 2022-10-31
> Did you see my question?

My bad, hadn't noticed you had already replied by the time I finished commenting.

> My concern is that by allowing one set of trees to become unbalanced (hence potentially deep for their size), you could break those proofs (and bounds).

Right. The idea here is that the act of balancing that `link` performs pays a cost anyway since it needs to navigate the unbalanced bits to balance them but we don't get to amortize that cost since we only do one operation on the tree and then throw it away. A bit like linear scanning of an array being faster than building a binomial queue in O(n) but then doing a single lookup.

That said, I don't currently have a proof, only a gut feeling backed by no more than a specific benchmark. I'll peek at the published proofs.